Add a constants file
The string enums can optionally be used in TypeScript projects in place
of the original constant values for better type inference.

The bitfield enums can also be used in contexts where there is actually no WebGPU
implementation but those values are needed (e.g. in Node when generating
test lists in the CTS).