Update to latest spec (#21)
Note this gets ahead of implementations by removing setSubData. However otherwise this should be compatible at least with current Canary (except for the many unimplemented parts).

Also takes an intermediate state in some of the ongoing spec changes (in particular mapAsync) but we'll fix that up when we have changes in the spec.