Roll gpuweb 7716424b -> 26e16be2b
This roll also switches the bikeshed-to-ts dependency to a
personal fork (ideally temporarily) in order to resolve issues
with the use of namespaces in the spec when calling `npm generate`.
This will unblock future rolls.

The generator changes did cause some of the bitflag definitions to
be shuffled to the end of the generated document, so this PR moves
them in order to make future rolls easier.