Run tsdoc on PRs, move build to a subdir, fix warnings
The tsdoc build broke recently because of the inclusion of VideoFrame.
This fixes that, and prevents it from happening again. It also moves the
build to a subdirectory cleans up the dependencies for more tests in the
next PR.

I had to replace the TypeDoc API usage with a command-line invocation
because I can't figure out how to enable treatWarningsAsErrors and then
exit>0 in the API usage.