Generate by using diff apply
In this version of generate.mjs we assume there is a file ./generated/index-temp.d.ts
that **was** created by calling bikeshed, applying our patches, and calling prettier
in the previous commit. This should give us the diff needed to take our generated file
to the current dist state.

We then generate a current bikeshed-fixup-prettier generated/index-temp.d.ts.
We copy that to generated/index.d.ts and apply the previous patch. If it applies
it should have just the new changes from index.bs